% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cake_recipe.R
\name{cake_recipe}
\alias{cake_recipe}
\title{Cake Recipe Conversion}
\usage{
cake_recipe(guest_count, flavor)
}
\arguments{
\item{guest_count}{number of people eating the cake}

\item{flavor}{choice of flavor}
}
\value{
Returns the cake recipe based on number of people
}
\description{
Cake Recipe Conversion
}
