% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empanadas.R
\name{empanadas}
\alias{empanadas}
\title{Calculate cost of making empanadas}
\usage{
empanadas(price, time)
}
\arguments{
\item{price}{price per empanada}

\item{time}{how long it took to make the emanada, in minutes}
}
\description{
Calculate cost of making empanadas
}
